Add files via upload
Edited some warnings with the help of PyCharm and added functionality to end the game (and not have the bot in an infinite loop of random number generation).  Most of these edits are changes to object names to make them more "pythonic" as well as some very minor formatting changes.