New and improved bot!
The previous approach was WAAAAY too brute force. Updated bot to use a new base strategy and some smart checking to reduce edge-cases down to three unique scenarios (as well as factoring in rotations/reflections of those scenarios).  I have not been this confident since the last time I asked @MicahWW to check the bot, and that request is going out again now!