Checkers Begins
The very beginning of a checkers module!  Includes *some* backend (just enough to proof-of-concept run it) and *some* terminal features (ditto on the "just enough").

Still needs double-jump, "king-ing", a check to prevent un-king-ed pieces from moving backwards (so maybe king-ing DOESN'T need to be added? 🤔) end-of-game winner messages, option for obligatory jumping... I'm sure I'll think of more.

This commit is mostly meant to allow @MicahWW to give input and suggest changes while this is still in the relatively early stages.  I suspect the col,row format will be a contentious discussion, I do already have ideas for where an earlier switch-around for those could go, but we'll pin down the details.  Other talking points include:
 - player numbering (zero-indexing players in TicTacToe is fine but for Checkers this is a hill I will die on)
 - the placement of jumpPiece in checkValidMove (I don't love it, open to conversation)
 - the colors of pieces
 - if out-of-bounds spaces should be colored differently than in-bounds spaces (that would take a bit of doing and would be good to start discussing now)
 - invalidEntryMessage as basically a variable/constant

Lastly... can I do a Batman titles thing for the next two commits on this topic?  So my next one would be "Checkers - Dark Knight" or something like that