Block double-middle strategy
Shortly after the most recent round of bot improvements, a strategy was discovered that resulted in winning against the bot 30-50% of the time.  The bot has been upgraded to defend against this strategy, while maintaining that sense of unpredictability we all know and love. Minor updates to some comments while I was in the neighborhood.

This also uncovered a bug with TicTacToe.move_history, specifically, the move history was retained over the course of several games, causing other bot bugs and would have potentially caused crashes with planned future features (like undo move).  This was fixed by adding a line to resetGame that resets the move history along with the board and game state.

A test was added to pyTest to ensure the bot continues to avoid the trap that started this whole thing in the first place (it looks complicated but takes all of 0.03 seconds to run and adds 0.00 seconds to running the entire test suite).

I feel like I did a pretty good job staying on task with this one, all of those things were fairly closely related! :D